chore: try moving the test run to a separate script
